Add suitable and not suitable training widgets for weather conditions
